requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.14.1
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
appdirs==1.4.4
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.1.1
Babel==2.15.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
bowler==0.9.0
cachelib==0.9.0
cachetools==5.3.3
cattrs==23.2.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.1
click-params==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
configobj==5.0.8
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
cycler==0.12.1
dask==2024.6.2
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==6.1.3
docopt==0.6.2
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.3
dvc-data==3.15.1
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.4.0
dynaconf==3.2.5
email-validator==1.3.1
entrypoints==0.4
evidently==0.4.23
executing==2.0.1
Faker==26.0.0
fastapi==0.99.1
fastapi-utils==0.2.1
fastavro==1.9.4
fastjsonschema==2.20.0
feast==0.35.0
filelock==3.15.4
fissix==24.4.24
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
flask-session2==1.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gast==0.6.0
gensim==4.3.2
giskard==2.14.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
google-re2==1.1.20240601
googleapis-common-protos==1.63.2
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.3
great-expectations==0.18.17
greenlet==3.0.3
griffe==0.47.0
grpcio==1.64.1
grpcio-health-checking==1.62.2
grpcio-reflection==1.62.2
grpcio-tools==1.62.2
gto==1.7.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.19.1
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
ipinfo==5.0.1
ipykernel==6.29.4
ipython==8.26.0
ipywidgets==8.1.3
isodate==0.6.1
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle==1.6.14
keras==3.4.1
kiwisolver==1.4.5
kombu==5.3.7
langdetect==1.0.9
lazy-object-proxy==1.10.0
libclang==18.1.1
lightgbm==4.4.0
limits==3.13.0
linkify-it-py==2.0.3
litestar==2.9.1
locket==1.0.0
lockfile==0.12.2
makefun==1.15.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mixpanel==4.10.1
ml-dtypes==0.3.2
mlflow==2.14.1
mlflow-skinny==2.14.1
mmh3==4.1.0
more-itertools==10.3.0
moreorless==0.4.0
msgspec==0.18.6
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
mypy-protobuf==3.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
num2words==0.5.13
numpy==1.24.4
oauth2client==4.1.3
omegaconf==2.3.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandavro==1.5.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pathspec==0.12.1
patsy==0.5.6
pendulum==2.1.2
pexpect==4.9.0
pillow==10.3.0
platformdirs==3.11.0
plotly==5.22.0
pluggy==1.5.0
polyfactory==2.16.0
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.23.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.18.4
pydot==1.4.2
PyDrive2==1.19.0
pygit2==1.15.0
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.7.0
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==2.4.7
pytest==8.2.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-argparse==1.5.2
rich-click==1.8.3
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.17.17
scikit-learn==1.5.0
scipy==1.11.4
scmrepo==3.3.5
seaborn==0.13.2
secure==0.3.0
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.7.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
shyaml==0.6.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.41
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
sqlparse==0.5.0
sqltrie==0.11.0
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.2
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==2.13.3
typer==0.12.3
types-protobuf==5.27.0.20240626
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validators==0.18.2
vine==5.1.0
volatile==2.1.0
voluptuous==0.15.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.11
wrapt==1.16.0
WTForms==3.0.1
xxhash==3.4.1
yarl==1.9.4
zc.lockfile==3.0.post1
zenml==0.58.2
zipp==3.19.2
zstandard==0.22.0
gradio==3.50.2